Spec 743.3 — CIA timer (ciat) + TOD absolute clocks monotonic
Fidelity-critical slice. Classified absolute-clk vs hardware-register per site
(read VICE-port intent first, no mechanical u32 removal):

- ciat.ts: setAlarm predict-walk — CLOCK_MAX 0xffffffff → CLOCK_NEVER, drop >>>0
  on aclk/tmp (absolute fire clock). update()'s this.clk was already monotonic.
  16-bit cnt/latch keep & 0xffff.
- cia6526-vice.ts: rclk = clk - offset (drop u32, 9 sites); idle/sdr alarm clks
  clk+N (drop u32); ta_alarmclk/tb_alarmclk + ciatSetAlarm sentinel → CLOCK_NEVER;
  ifr_clock+1, rdi===rclk-1, ciacoreIntta/Inttb offset, runPendingAlarms dispatch
  clk → monotonic. KEPT u32 on the delay/sdr_delay IRQ-pipeline REGISTERS + the
  >>>0 on `delay`. Updated the stale early-boot `u32(0-1)` comment (rclk now
  negative-not-huge; realClk guard still protects).
- cia-tod.ts: the 3 absolute todclk = currentClk + ticks (drop u32). todticks /
  ticks_per_sec / power_ticks keep u32 (small periods needing floor, < 2^32).
- alarm-context.ts: alarmContextTimeWarp de-u32 (dead code, kept correct).

After this slice there is NO u32/>>>0 on an absolute maincpu clk anywhere; the
remaining u32 are register widths (delay/sdr_delay), small deltas/lengths
(lastIFlagClearInstrLen, dispatch offset) and TOD periods.

Regression (CIA timing is fidelity-critical): cia-suite 16/16, cia-fidelity 22/22,
probe:743-3 11/11 (ciat underflow + TOD tick monotonic past 2^32, TOD interval
19704≈985248/50, counter still 16-bit). Games verified visually by reading the
PNGs: motm (riverboat + 3 menu lines), MM (character select), IM2 (man in elevator)
— LNR/Scramble/Pawn/Polarbear re-rendering. Below-2^32 behaviour is identical
(u32 on <2^32 = no-op).

+++ src/runtime/headless/alarm/alarm-context.ts
@@ -217,8 +217,10 @@ export function alarmContextRestoreSchedule(
  *
  * Shifts every pending alarm's clk by `warp_amount` in the given
  * direction. Direction 0 → no-op (matches VICE early-return). Positive
- * → add, negative → subtract. We fold through `u32` so wrap matches
- * VICE uint32 semantics.
+ * → add, negative → subtract. Spec 743: absolute clk is monotonic, so the
+ * shift is a plain add/subtract (no u32 wrap). NOTE: currently has no caller
+ * (the monotonic CLOCK never needs a clkguard-style rebase); kept correct in
+ * case a future rewind/offset feature wires it.
  */
 export function alarmContextTimeWarp(
   context: AlarmContext,
@@ -232,20 +234,16 @@ export function alarmContextTimeWarp(
   for (let i = 0; i < context.num_pending_alarms; i++) {
     const slot = context.pending_alarms[i]!;
     if (warpDirection > 0) {
-      slot.clk = u32(slot.clk + warpAmount);
+      slot.clk = slot.clk + warpAmount;
     } else {
-      slot.clk = u32(slot.clk - warpAmount);
+      slot.clk = slot.clk - warpAmount;
     }
   }
 
   if (warpDirection > 0) {
-    context.next_pending_alarm_clk = u32(
-      context.next_pending_alarm_clk + warpAmount,
-    );
+    context.next_pending_alarm_clk = context.next_pending_alarm_clk + warpAmount;
   } else {
-    context.next_pending_alarm_clk = u32(
-      context.next_pending_alarm_clk - warpAmount,
-    );
+    context.next_pending_alarm_clk = context.next_pending_alarm_clk - warpAmount;
   }
 }
